1.1.8.2-Leisure Upgrade
- Fix bug in GetDimensionalBalance affecting POG available coins
- Modify POG coin age calculation to use greater precision

+++ src/primitives/block.h
@@ -87,7 +87,7 @@ public:
 	int Height;
 	int HeightLast;
 	uint256 LastBlockHash;
-	int64_t Age;
+	double Age;
 	std::string TXID;
 	int Ordinal;
